refactor(services): move to new variable-list component
