+++ libs/pages/application/src/lib/feature/page-variables-feature/page-variables-feature.tsx
@@ -1,21 +1,25 @@
 import { APIVariableScopeEnum } from 'qovery-typescript-axios'
-import { useContext, useEffect, useState } from 'react'
+import { useContext, useEffect } from 'react'
 import { useParams } from 'react-router-dom'
 import { match } from 'ts-pattern'
+import { useDeployEnvironment } from '@qovery/domains/environments/feature'
 import { useService } from '@qovery/domains/services/feature'
-import { VariableList, useVariables } from '@qovery/domains/variables/feature'
-import { environmentVariableFactoryMock } from '@qovery/shared/factories'
-import { type EnvironmentVariableSecretOrPublic } from '@qovery/shared/interfaces'
-import { type TableHeadProps } from '@qovery/shared/ui'
+import { VariableList, useDeleteVariable } from '@qovery/domains/variables/feature'
+import { DEPLOYMENT_LOGS_URL, ENVIRONMENT_LOGS_URL } from '@qovery/shared/routes'
+import { ToastEnum, toast, useModal, useModalConfirmation } from '@qovery/shared/ui'
 import { useDocumentTitle } from '@qovery/shared/util-hooks'
+import { environmentVariableFile } from '@qovery/shared/util-js'
 import { ApplicationContext } from '../../ui/container/container'
-import PageVariables from '../../ui/page-variables/page-variables'
-
-const placeholder = environmentVariableFactoryMock(5) as EnvironmentVariableSecretOrPublic[]
+import {
+  CrudEnvironmentVariableModalFeature,
+  EnvironmentVariableCrudMode,
+} from '../crud-environment-variable-modal-feature/crud-environment-variable-modal-feature'
 
 export function PageVariablesFeature() {
   useDocumentTitle('Environment Variables – Qovery')
-  const { environmentId, applicationId = '' } = useParams()
+  const { environmentId = '', organizationId = '', projectId = '', applicationId = '' } = useParams()
+  const { openModal, closeModal } = useModal()
+  const { openModalConfirmation } = useModalConfirmation()
 
   const { data: service } = useService({
     environmentId,
@@ -29,98 +33,90 @@ export function PageVariablesFeature() {
     .with('HELM', () => APIVariableScopeEnum.HELM)
     .otherwise(() => undefined)
 
-  const { data: sortVariableMemo = [], isLoading } = useVariables({
-    parentId: applicationId,
-    scope,
+  const { mutateAsync: deleteVariable } = useDeleteVariable()
+  const { mutate: actionRedeployEnvironment } = useDeployEnvironment({
+    projectId,
+    logsLink: ENVIRONMENT_LOGS_URL(organizationId, projectId, environmentId) + DEPLOYMENT_LOGS_URL(applicationId),
   })
 
   const serviceType = service?.serviceType
 
-  const [data, setData] = useState<EnvironmentVariableSecretOrPublic[]>(sortVariableMemo || placeholder)
-
-  const { setShowHideAllEnvironmentVariablesValues } = useContext(ApplicationContext)
+  const { showHideAllEnvironmentVariablesValues, setShowHideAllEnvironmentVariablesValues } =
+    useContext(ApplicationContext)
 
   useEffect(() => {
     setShowHideAllEnvironmentVariablesValues(false)
   }, [applicationId, serviceType])
 
-  useEffect(() => {
-    if (isLoading) {
-      setData(placeholder)
-    } else {
-      // XXX: This should be done using `useMutationState` in tanstack-query v5 (we are currently still in v4)
-      // https://tanstack.com/query/v5/docs/react/guides/optimistic-updates
-      setData((previousData) =>
-        previousData.length === 0 || previousData === placeholder
-          ? sortVariableMemo
-          : sortVariableMemo.map((variable) => ({
-              ...variable,
-              is_new: !previousData.find((v) => v.key === variable.key),
-            }))
-      )
-    }
-  }, [sortVariableMemo, isLoading])
-
-  const tableHead: TableHeadProps<EnvironmentVariableSecretOrPublic>[] = [
-    {
-      title: !isLoading ? `${data?.length} variable${data?.length && data.length > 1 ? 's' : ''}` : `0 variable`,
-      className: 'px-4 py-2',
-    },
-    {
-      title: 'Update',
-      className: 'pl-4 pr-12 text-end',
-    },
-    {
-      title: 'Value',
-      className: 'px-4 py-2 border-b-neutral-200 border-l h-full',
-      filter: [
-        {
-          title: 'Sort by privacy',
-          key: 'variable_kind',
-        },
-      ],
-    },
-    {
-      title: 'Service link',
-      filter: [
-        {
-          title: 'Sort by service',
-          key: 'service_name',
-        },
-      ],
-    },
-    {
-      title: 'Scope',
-      filter: [
-        {
-          title: 'Sort by scope',
-          key: 'scope',
-        },
-      ],
-    },
-  ]
-
   return (
-    <>
-      <PageVariables
-        key={data.length}
-        tableHead={tableHead}
-        variables={
-          !isLoading
-            ? data.map((variable) => ({
-                ...variable,
-                // XXX: this is needed to comply with the current table implementation.
-                // It should be removed when migrating to tanstack-table
-                variable_kind: variable.is_secret ? ('secret' as const) : ('public' as const),
-              }))
-            : placeholder
-        }
-        setData={setData}
-        isLoading={isLoading}
-        serviceType={serviceType}
-      />
-      {scope && <VariableList currentScope={scope} parentId={applicationId} className="bg-white" />}
-    </>
+    <div className="mt-2 bg-white rounded-sm">
+      {scope && (
+        <VariableList
+          showAll={showHideAllEnvironmentVariablesValues}
+          currentScope={scope}
+          parentId={applicationId}
+          onCreateVariable={(variable, variableType) => {
+            openModal({
+              content: (
+                <CrudEnvironmentVariableModalFeature
+                  closeModal={closeModal}
+                  variable={variable}
+                  type={variableType}
+                  mode={EnvironmentVariableCrudMode.CREATION}
+                  organizationId={organizationId}
+                  applicationId={applicationId}
+                  projectId={projectId}
+                  environmentId={environmentId}
+                  serviceType={service?.serviceType}
+                  isFile={environmentVariableFile(variable)}
+                />
+              ),
+            })
+          }}
+          onEditVariable={(variable) => {
+            openModal({
+              content: (
+                <CrudEnvironmentVariableModalFeature
+                  closeModal={closeModal}
+                  variable={variable}
+                  mode={EnvironmentVariableCrudMode.EDITION}
+                  organizationId={organizationId}
+                  applicationId={applicationId}
+                  projectId={projectId}
+                  environmentId={environmentId}
+                  type={variable.variable_type}
+                  serviceType={service?.serviceType}
+                  isFile={environmentVariableFile(variable)}
+                />
+              ),
+            })
+          }}
+          onDeleteVariable={(variable) => {
+            openModalConfirmation({
+              title: 'Delete variable',
+              name: variable.key,
+              isDelete: true,
+              action: async () => {
+                await deleteVariable({ variableId: variable.id })
+                let name = variable.key
+                if (name && name.length > 30) {
+                  name = name.substring(0, 30) + '...'
+                }
+                const toasterCallback = () => actionRedeployEnvironment({ environmentId })
+                toast(
+                  ToastEnum.SUCCESS,
+                  'Deletion success',
+                  `${name} has been deleted. You need to redeploy your environment for your changes to be applied.`,
+                  toasterCallback,
+                  undefined,
+                  'Redeploy'
+                )
+              },
+            })
+          }}
+        />
+      )}
+    </div>
   )
 }
