--- libs/pages/application/src/lib/feature/table-row-environment-variable-feature/table-row-environment-variable-feature.spec.tsx
@@ -1,19 +1,0 @@
-import { render } from '__tests__/utils/setup-jest'
-import { mockSecretEnvironmentVariable } from '@qovery/shared/factories'
-import TableRowEnvironmentVariableFeature, {
-  type TableRowEnvironmentVariableFeatureProps,
-} from './table-row-environment-variable-feature'
-
-const props: TableRowEnvironmentVariableFeatureProps = {
-  variable: mockSecretEnvironmentVariable(false, false),
-  dataHead: [],
-  columnsWidth: '',
-  isLoading: false,
-}
-
-describe('TableRowEnvironmentVariableFeature', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<TableRowEnvironmentVariableFeature {...props} />)
-    expect(baseElement).toBeTruthy()
-  })
-})
