--- libs/pages/application/src/lib/ui/page-variables/page-variables.spec.tsx
@@ -1,18 +1,0 @@
-import { render } from '__tests__/utils/setup-jest'
-import { secretEnvironmentVariableFactoryMock } from '@qovery/shared/factories'
-import PageVariables, { type PageVariablesProps } from './page-variables'
-
-const props: PageVariablesProps = {
-  variables: secretEnvironmentVariableFactoryMock(3, false, false),
-  filterData: secretEnvironmentVariableFactoryMock(3, false, false),
-  listHelpfulLinks: [],
-  setFilterData: jest.fn(),
-  tableHead: [],
-}
-
-describe('PageVariables', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<PageVariables {...props} />)
-    expect(baseElement).toBeTruthy()
-  })
-})
